# v3.5.2: Optimize network closing and Improve file canReadMethod