--- a/WaarpFtp/core/src/main/java/org/waarp/ftp/core/command/FtpCommandCode.java
+++ b/WaarpFtp/core/src/main/java/org/waarp/ftp/core/command/FtpCommandCode.java
@@ -49,7 +49,8 @@ public enum FtpCommandCode {
    */
   Connection(ConnectionCommand.class, null,
              org.waarp.ftp.core.command.access.USER.class,
-             org.waarp.ftp.core.command.rfc2389.FEAT.class),
+             org.waarp.ftp.core.command.rfc2389.FEAT.class,
+             org.waarp.ftp.core.command.rfc2389.OPTS.class),
   // XXX ACCESS CONTROL COMMAND
   /**
    * The argument field is a Telnet string identifying the user. The user