--- a/WaarpR66/src/main/java/org/waarp/openr66/commander/Commander.java
+++ b/WaarpR66/src/main/java/org/waarp/openr66/commander/Commander.java
@@ -64,6 +64,7 @@ public class Commander implements CommanderInterface {
   private static final WaarpLogger logger =
       WaarpLoggerFactory.getLogger(Commander.class);
 
+  public static final int LIMIT_MAX_SUBMIT = 10000;
   public static final int LIMIT_SUBMIT = 1000;
 
   private InternalRunner internalRunner;