--- a/WaarpR66/src/main/java/org/waarp/openr66/protocol/localhandler/ServerActions.java
+++ b/WaarpR66/src/main/java/org/waarp/openr66/protocol/localhandler/ServerActions.java
@@ -1258,8 +1258,13 @@ public class ServerActions extends ConnectionActions {
     }
     final NetworkChannelReference ncr =
         localChannelReference.getNetworkChannelObject();
-    NetworkTransaction.shuttingDownNetworkChannel(ncr);
-    NetworkTransaction.shuttingdownNetworkChannelsPerHostID(ncr.getHostId());
+    ncr.lockNetwork();
+    long time = ncr.shutdownAllowed();
+    if (time == 0) {
+      logger.info("Will close networkChannel {}", ncr.nbLocalChannels());
+      NetworkTransaction.shuttingDownNetworkChannel(ncr);
+      NetworkTransaction.shuttingdownNetworkChannelsPerHostID(ncr.getHostId());
+    }
   }
 
   /**