… WaarpR66/pom.xml
… WaarpR66/src/main/java/org/waarp/openr66/protocol/utils/ChannelUtils.java
… WaarpR66/src/main/java/org/waarp/openr66/protocol/utils/FileUtils.java
… WaarpR66/src/test/java/org/waarp/openr66/protocol/it/ScenarioBaseLoop.java
… WaarpR66/src/test/java/org/waarp/openr66/protocol/it/ScenarioBaseLoopBenchmark.java
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/OpenR66-authent.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R1/conf/ruleloop.rule.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R1/conf/rulerecv-any.rule.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R1/conf/rulerecvthrough-any.rule.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R1/conf/rulesend-any.rule.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R1/conf/server_1.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R1/conf/server_1_SQLDB.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R2/conf/ruleloop.rule.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R2/conf/rulerecv-any.rule.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R2/conf/rulerecvthrough-any.rule.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R2/conf/rulesend-any.rule.xml
… WaarpR66/src/test/resources/it/scenario_loop_benchmark/R2/conf/server_2.xml
… doc/waarp-r66/source/changes.rst
… doc/waarp-r66/source/conf.py
… pom.xml